+++ navigation/AppNavigator.js
@@ -1,0 +1,8 @@
+import React from "react";
+import UserScreen from "../screens/UserScreen";
+
+const AppNavigator = () => {
+  return <UserScreen />;
+};
+
+export default AppNavigator;
